requirements.txt
telethon
pymongo
requests
python-dotenv